regen stubs
